+++ src/main/kotlin/io/kjson/uri/ExpressionElement.kt
@@ -34,8 +34,8 @@ class ExpressionElement(
     val prefix: Char?,
     val separator: Char,
     val reservedEncoding: Boolean,
-    val addVariableNames: Boolean = false,
-    val formsStyleEqualsSign: Boolean = false,
+    val addVariableNames: Boolean,
+    val formsStyleEqualsSign: Boolean,
 ) : Element {
 
     override fun appendTo(a: Appendable) {
@@ -45,19 +45,47 @@ class ExpressionElement(
             variable.value?.takeUnless {
                 it is List<*> && it.isEmpty() || it is Map<*, *> && it.isEmpty()
             }?.let { value ->
-                if (continuation)
-                    a.append(separator)
-                else
-                    prefix?.let { a.append(it) }
-                if (addVariableNames) {
-                    a.append(variable.name)
-                    if (formsStyleEqualsSign || !value.isEmptyOrAllNull())
-                        a.append('=')
+                if (reference.explode) {
+                    when (value) {
+                        is List<*> -> for (item in value) {
+                            item?.let { appendValue(a, variable.name, it, null, addVariableNames, continuation) }
+                            continuation = true
+                        }
+                        is Map<*, *> -> {
+                            for (entry in value.entries) {
+                                entry.value?.let { appendValue(a, entry.key.toString(), it, null, true, continuation) }
+                                continuation = true
+                            }
+                        }
+                        else -> appendValue(a, variable.name, value, reference.characterLimit, addVariableNames,
+                                continuation)
+                    }
                 }
-                appendToString(a, value)
+                else
+                    appendValue(a, variable.name, value, reference.characterLimit, addVariableNames, continuation)
                 continuation = true
             }
         }
+    }
+
+    private fun appendValue(
+        a: Appendable,
+        name: String,
+        value: Any?,
+        characterLimit: Int?,
+        addNames: Boolean,
+        continuation: Boolean,
+    ) {
+        if (continuation)
+            a.append(separator)
+        else
+            prefix?.let { a.append(it) }
+        if (addNames) {
+            a.append(name)
+            if (formsStyleEqualsSign || value != null && !value.isEmptyOrAllNull())
+                a.append('=')
+        }
+        value?.let { appendToString(a, it, characterLimit) }
     }
 
     private fun Any.isEmptyOrAllNull(): Boolean = when (this) {
@@ -67,16 +95,16 @@ class ExpressionElement(
         else -> toString().isEmpty()
     }
 
-    private fun appendToString(a: Appendable, obj: Any) {
+    private fun appendToString(a: Appendable, obj: Any, characterLimit: Int?) {
         when (obj) {
-            is String -> a.append(obj.encodeUTF8().encode())
+            is String -> a.append(obj.applyCharacterLimit(characterLimit).encodeUTF8().encode())
             is List<*> -> {
                 var continuation = false
                 for (item in obj) {
                     if (item != null) {
                         if (continuation)
                             a.append(',')
-                        appendToString(a, item)
+                        appendToString(a, item, null)
                         continuation = true
                     }
                 }
@@ -87,19 +115,32 @@ class ExpressionElement(
                     entry.key?.let {
                         if (continuation)
                             a.append(',')
-                        appendToString(a, it)
+                        appendToString(a, it, null)
                         continuation = true
                     }
                     entry.value?.let {
                         if (continuation)
                             a.append(',')
-                        appendToString(a, it)
+                        appendToString(a, it, null)
                         continuation = true
                     }
                 }
             }
             else -> a.append(obj.toString().encodeUTF8().encode())
         }
+    }
+
+    private fun String.applyCharacterLimit(characterLimit: Int?): String {
+        if (characterLimit == null || length <= characterLimit)
+            return this
+        var size = 0
+        for (i in 0 until characterLimit) {
+            if (size + 1 < length && this[size].isHighSurrogate() && this[size + 1].isLowSurrogate())
+                size += 2
+            else
+                size++
+        }
+        return if (size == length) this else substring(0, size)
     }
 
     private fun String.encode(): String = if (reservedEncoding) encodeReserved() else encodeSimple()